rename tab-notification utility to browser-tab-notification
