+++ core-libs/core/src/util/browser-tab-notification/index.ts
@@ -4,4 +4,4 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export * from './abstract-tab-notification.service';
+export * from './abstract-browser-tab-notification.service';
